+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changelog for NeoFS Node
 - Process designation event of the mainnet RoleManagement contract (#3134)
 - Storage nodes running out of GAS because `putContainerSize` was not paid for by proxy (#3167)
 - Write-cache flushing loop to drop objects (#3169)
+- FS chain client pool overflows and event ordering (#3163)
 
 ### Changed
 - Number of cuncurrenly handled notifications from the chain was increased from 10 to 300 for IR (#3068)